add query_jobs_admin and  tests - this mirrors ee2 (and thus job browser bff); and it helps in us- and test-ability, etc. - also started work on separating tests into different files, per method - adjusted the job state property (only used in job browser 2, and in that case will only be used for a few weeks)
