+++ lib/kb_Metrics/kb_MetricsServer.py
@@ -350,6 +350,10 @@ class Application(object):
                              name='kb_Metrics.query_jobs',
                              types=[dict])
         self.method_authentication['kb_Metrics.query_jobs'] = 'required'  # noqa
+        self.rpc_service.add(impl_kb_Metrics.query_jobs_admin,
+                             name='kb_Metrics.query_jobs_admin',
+                             types=[dict])
+        self.method_authentication['kb_Metrics.query_jobs_admin'] = 'required'  # noqa
         self.rpc_service.add(impl_kb_Metrics.get_job,
                              name='kb_Metrics.get_job',
                              types=[dict])
