+++ test/db_files/ci_userjobstate.jobstate.json
@@ -605,5 +605,46 @@
 			"started": 1516822526990,
 			"complete": true,
 			"error": false
+		},
+		{
+			"_id": "5d4493a9aa5a4d298c5dc940",
+			"user": "eapearson",
+			"authstrat": "kbaseworkspace",
+			"authparam": "15206",
+			"created": 1500004550060,
+			"updated": 1500004584594,
+			"estcompl": null,
+			"status": "canceled by user",
+			"desc": "Execution engine job for kb_deseq.run_deseq2_app",
+			"started": 1500004554670,
+			"complete": true,
+			"error": false
+		},
+		{
+			"_id": "5d4493a9aa5a4d298c5dc950",
+			"user": "eapearson",
+			"authstrat": "kbaseworkspace",
+			"authparam": "15206",
+			"created": 1500004550060,
+			"updated": 1500004584594,
+			"estcompl": null,
+			"status": "",
+			"desc": "Execution engine job for kb_deseq.run_deseq2_app",
+			"started": 1500004554670,
+			"complete": false,
+			"error": false
+		},
+		{
+			"_id": "5d4493a9aa5a4d298c5dc960",
+			"user": "eapearson",
+			"authstrat": "kbaseworkspace",
+			"authparam": "15206",
+			"created": 1500004550060,
+			"updated": 1500004584594,
+			"estcompl": null,
+			"status": "",
+			"desc": "A queued job",
+			"complete": false,
+			"error": false
 		}
 	]
