+++ test/disabled/kb_Metrics_server_query_jobs_admin_test.py
@@ -26,23 +26,55 @@ from kb_Metrics.metricsdb_controller import MetricsMongoDBController
 from kb_Metrics.NarrativeCache import NarrativeCache
 from kb_Metrics.Test import Test
 
-TOTAL_COUNT = 9
+debug = False
 
-class kb_Metrics_query_jobs_Test(Test):
+TOTAL_COUNT = 46
+
+def print_debug(msg):
+    if not debug:
+        return
+    t = str(datetime.datetime.now())
+    print ("{}:{}".format(t, msg))
+    
 
 
-    # def mock_MongoMetricsDBI(self, mongo_host, mongo_dbs,
-    #                          mongo_user, mongo_psswd):
-    #     self.mongo_clients = dict()
-    #     self.metricsDBs = dict()
-    #     for m_db in mongo_dbs:
-    #         self.mongo_clients[m_db] = MongoClient()
-    #         self.metricsDBs[m_db] = self.mongo_clients[m_db][m_db]
+    def getWsClient(self):
+        return self.__class__.wsClient
+
+    def getWsName(self):
+        if hasattr(self.__class__, 'wsName'):
+            return self.__class__.wsName
+        suffix = int(time.time() * 1000)
+        wsName = "test_kb_Metrics_" + str(suffix)
+        ret = self.getWsClient().create_workspace({'workspace': wsName})  # noqa
+        self.__class__.wsName = wsName
+        return wsName
+
+    def getImpl(self):
+        return self.__class__.serviceImpl
+
+    def getContext(self):
+        return self.__class__.ctx
+
+    def mock_MongoMetricsDBI(self, mongo_host, mongo_dbs,
+                                mongo_user, mongo_psswd):
+        self.mongo_clients = dict()
+        self.metricsDBs = dict()
+        for m_db in mongo_dbs:
+            self.mongo_clients[m_db] = MongoClient()
+            self.metricsDBs[m_db] = self.mongo_clients[m_db][m_db]
+
+
+class kb_Metrics_query_jobs_admin_Test(Test):
+
+    # def __init__(self, test):
+    #     print('here?')
+    #     super().__init__()
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_no_params")
     def test_run_MetricsImpl_query_jobs_no_params(self):
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
         })
 
         self.assertEqual(len(ret), 1)
@@ -52,13 +84,12 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
         self.assertEqual(result['found_count'], TOTAL_COUNT)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_max_time_range")
     def test_run_MetricsImpl_query_jobs_max_time_range(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'epoch_range': [0, now]
         })
 
@@ -69,13 +100,12 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
         self.assertEqual(result['found_count'], TOTAL_COUNT)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_max_time_range_paged")
     def test_run_MetricsImpl_query_jobs_max_time_range_paged(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'epoch_range': [0, now],
             'offset': 5,
             'limit': 5
@@ -88,14 +118,13 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
         self.assertEqual(result['found_count'], TOTAL_COUNT)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
-        self.assertEqual(len(result['job_states']), 4)
+        self.assertEqual(len(result['job_states']), 5)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_user")
     def test_run_MetricsImpl_query_jobs_one_user(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'epoch_range': [0, now],
             'filter': {
                 'user_id': ['psdehal']
@@ -108,16 +137,15 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIsInstance(result, dict) 
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
-        self.assertEqual(result['found_count'], 0)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
+        self.assertEqual(result['found_count'], 1)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_job")
     def test_run_MetricsImpl_query_jobs_one_job(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
-                'job_id': ['5d4493a9aa5a4d298c5dc940']
+                'job_id': ['5a68dffce4b0ace8f870f586']
             }
         })
 
@@ -128,13 +156,12 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
         self.assertEqual(result['found_count'], 1)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_job")
     def test_run_MetricsImpl_query_jobs_filter_by_status_queue(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
                 'status': ['queue']
             }
@@ -145,14 +172,13 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIsInstance(result, dict) 
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
-        self.assertEqual(result['found_count'], 1)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
+        self.assertEqual(result['found_count'], 2)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_job")
     def test_run_MetricsImpl_query_jobs_filter_by_status_run(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
                 'status': ['run']
             }
@@ -164,13 +190,12 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
         self.assertEqual(result['found_count'], 2)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_job")
     def test_run_MetricsImpl_query_jobs_filter_by_status_complete(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
                 'status': ['complete']
             }
@@ -182,14 +207,13 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIsInstance(result, dict) 
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
-        self.assertEqual(result['found_count'], 2)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
+        self.assertEqual(result['found_count'], 29)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_job")
     def test_run_MetricsImpl_query_jobs_filter_by_status_error(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
                 'status': ['error']
             }
@@ -201,14 +225,13 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIsInstance(result, dict) 
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
-        self.assertEqual(result['found_count'], 5)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
+        self.assertEqual(result['found_count'], 13)
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_job")
     def test_run_MetricsImpl_query_jobs_filter_by_status_cancel(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
                 'status': ['cancel']
             }
@@ -220,15 +243,14 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIsInstance(result, dict) 
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
-        self.assertEqual(result['found_count'], 1)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
+        self.assertEqual(result['found_count'], 2)
 
 
     # Uncomment to skip this test
     # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_one_job")
     def test_run_MetricsImpl_query_jobs_filter_by_status_cancel_or_error(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
                 'status': ['cancel', 'error']
             }
@@ -240,15 +262,12 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIsInstance(result, dict) 
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
-        self.assertEqual(result['found_count'], 6)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
+        self.assertEqual(result['found_count'], 15)
 
 
-    # Uncomment to skip this test
-    # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_filter_by_status_queue_or_run")
     def test_run_MetricsImpl_query_jobs_filter_by_status_queue_or_run(self):
         now = int(round(time.time() * 1000))
-        ret = self.getImpl().query_jobs(self.getContext(), {
+        ret = self.getImpl().query_jobs_admin(self.getContext(), {
             'filter': {
                 'status': ['queue', 'run']
             }
@@ -260,18 +279,14 @@ class kb_Metrics_query_jobs_Test(Test):
         self.assertIsInstance(result, dict) 
         self.assertIn('job_states', result)
         self.assertIn('found_count', result)
-        self.assertEqual(result['found_count'], 3)
-        self.assertEqual(result['total_count'], TOTAL_COUNT)
+        self.assertEqual(result['found_count'], 4)
 
-    # Uncomment to skip this test
-    # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_sort2")
     def test_run_MetricsImpl_query_jobs_sort2(self):
-        from_time = 1546329600000 # 1/1/2019
-        to_time = 1572591600000 # 11/1/2019
+        total_count = TOTAL_COUNT
         test_data = [
-            # Can't test for updated -- it is not always populated!
-            # ['updated', 'modification_time', None, 1516822657338],
-            ['created', 'creation_time', 1500004550060, 1516822524141],
+            ['updated', 'finish_time', 1414192660701, 1516822657338],
+            ['created', 'finish_time', 1414192660701, 1516822657338],
+            ['user_id', 'user', 'arfath', 'wjriehl']
         ]
         data = []
         for field, expected_field, first, last in test_data:
@@ -287,9 +302,7 @@ class kb_Metrics_query_jobs_Test(Test):
                                 'field': field,
                                 'direction': direction
                             }
-                        ],
-                        'offset': 0,
-                        'limit': 100
+                        ]
                     },
                     'expected': [
                         {
@@ -304,20 +317,17 @@ class kb_Metrics_query_jobs_Test(Test):
                 })
 
         for datum in data:
-            ret = self.getImpl().query_jobs(self.getContext(), datum['input'])
+            ret = self.getImpl().query_jobs_admin(self.getContext(), datum['input'])
             self.assertEqual(len(ret), 1)
             self.assertIsInstance(ret[0], dict)
             result = ret[0]
             self.assertIsInstance(result, dict) 
             self.assertIn('job_states', result)
-            jobs = result['job_states']
-            # print("\n\nJOBS", jobs)
-
             self.assertIn('found_count', result)
             self.assertIn('total_count', result)
-            self.assertEqual(result['found_count'], TOTAL_COUNT)
-            self.assertEqual(result['total_count'], TOTAL_COUNT)
-            
+            self.assertEqual(result['found_count'], total_count)
+            self.assertEqual(result['total_count'], total_count)
+            jobs = result['job_states']
             for expected in datum['expected']:
                 item = expected['item']
                 if type(item) is str:
@@ -325,12 +335,9 @@ class kb_Metrics_query_jobs_Test(Test):
                         item = len(jobs) - 1
                 job = jobs[item]
                 for k, v in expected['fields'].items():
-                    self.assertEqual(job.get(k), v)
+                    self.assertEqual(job[k], v)
 
-    # Uncomment to skip this test
-    # @unittest.skip("skipped test_run_MetricsImpl_query_jobs_search")
     def test_run_MetricsImpl_query_jobs_search(self):
-       
         data = [
             # this one catches eapearson username
             {
@@ -343,7 +350,7 @@ class kb_Metrics_query_jobs_Test(Test):
                     ]
                 },
                 'expected': {
-                    'found_count': TOTAL_COUNT
+                    'found_count': 9
                 }
             },
             {
@@ -356,7 +363,7 @@ class kb_Metrics_query_jobs_Test(Test):
                     ]
                 },
                 'expected': {
-                    'found_count': TOTAL_COUNT
+                    'found_count': 9
                 }
             },
             # this one should catch one job by id
@@ -428,7 +435,7 @@ class kb_Metrics_query_jobs_Test(Test):
         ]
 
         for datum in data:
-            ret = self.getImpl().query_jobs(self.getContext(), datum['input'])
+            ret = self.getImpl().query_jobs_admin(self.getContext(), datum['input'])
             self.assertEqual(len(ret), 1)
             self.assertIsInstance(ret[0], dict)
             result = ret[0]
